Add files via upload
Data Source- HTTPS://RELATIONAL.FIT.CVUT.CZ/DATASET/FINANCIAL
Data warehousing is a powerful technique used by businesses, government agencies, and other organizations to store and analyze large amounts of data. One of the key aspects of designing a data warehouse is selecting an appropriate schema, and in this report, we present a star schema for a financial data warehouse, including a fact table and five dimension tables. We carefully chose the parameters for these tables by considering the available data from the financial dataset, which contains information about account holders, their transactions, loans, credit cards, and other relevant data.

Using a star schema reduces data redundancy, simplifies data retrieval, and improves query performance, making it an efficient and effective option for data warehousing. We chose this schema because it provides a simple way to retrieve data and can offer useful insights into customer behavior, loan payments, and account transactions. By analyzing this data, businesses and organizations can make informed decisions and improve their services.

The financial dataset of a bank was selected as the subject area for developing a proof-of-concept data warehouse/mart, as it offers a rich source of information that can be analyzed to improve services. By leveraging the insights provided by the data warehouse, bank managers can identify which customers are most profitable and which ones are not. Armed with this information, they can then take specific actions to improve services for their most valuable customers and minimize losses from unprofitable ones.

The project also offers important findings, including the use of SSIS and SSRS for constructing data warehouses and generating reports, the flexibility of Tableau reports in manipulating data, the suitability of graph databases for intricate relationship scenarios, and the rapid and effective CQL (Cypher Query Language) for data retrieval at higher levels of abstraction.

In summary, data warehousing is an essential tool for organizations looking to make informed decisions based on large amounts of data. By selecting an appropriate schema, such as the star schema we presented in this report, businesses can efficiently store and retrieve data, and gain valuable insights into customer behavior and preferences, ultimately leading to improved services and profitability.